fix merge & rearrange components to satisfy lint
